// app/models/customer.model.ts
export interface CustomerRequest {
    customerId: string;
    customerName: string;
    oneTime: boolean;
    active: boolean;
    lastName: string;
    firstName: string;
    middleName: string;
    prefix: string;
    companyName: string;
    printAs: string;
    primaryPhoneNo: string;
    secondaryPhoneNo: string;
    cellularPhoneNo: string;
    pagerNo: string;
    faxNo: string;
    primaryEmailAddress: string;
    secondaryEmailAddress: string;
    primaryUrl: string;
    secondaryUrl: string;
    addressLine1: string;
    addressLine2: string;
    city: string;
    stateProvince: string;
    zipPostalCode: string;
    country: string;
    isoCountryCode: string;
    excludedFromContactList: boolean;
    customerTypeId: string;
    salesRepEmployeeId: string;
    parentCustomerId: string;
    glGroupName: string;
    territoryId: string;
    attachmentsId: string;
    paymentTerm: string;
    offsetArGlAccountNo: string;
    defaultRevenueGlAccountNo: string;
    shippingMethod: string;
    resaleNumber: string;
    taxable: boolean;
    contactTaxGroupName: string;
    taxId: string;
    creditLimit: number;
    onHold: boolean;
    deliveryMethod: string;
    defaultInvoiceMessage: string;
    comments: string;
    defaultCurrency: string;
    printOptionArInvoiceTemplateName: string;
    printOptionOeQuoteTemplateName: string;
    printOptionOeOrderTemplateName: string;
    printOptionOeListTemplateName: string;
    printOptionOeInvoiceTemplateName: string;
    printOptionOeAdjustmentTemplateName: string;
    printOptionOeOtherTemplateName: string;
    primaryContactName: string;
    billToContactName: string;
    shipToContactName: string;
    restrictionType: string;
    restrictedLocations: string[];
    restrictedDepartments: string[];
}